tests: move unix-specific test to its own file
Other logmon tests should be portable.
